ml_project/requirements.txt
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.0
executing==1.2.0
Faker==15.1.3
fonttools==4.38.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-resources==5.10.0
iniconfig==1.1.1
ipykernel==6.16.2
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.4.4
jupyter-core==4.11.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
missingno==0.5.1
multimethod==1.9
nest-asyncio==1.5.6
networkx==2.8.7
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandas-profiling==3.4.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
pipeline==0.1.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
stack-data==0.6.0
statsmodels==0.13.2
sweetviz==2.1.4
tangled-up-in-unicode==0.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
visions==0.7.5
wcwidth==0.2.5
zipp==3.10.0
